+++ constants/products/product_schemas.py
@@ -1,0 +1,11 @@
+TYPE_SCHEMAS = {
+    'pokemon': {
+        'required': ['set', 'cards'],
+        'optional': ['sealed_products']
+    },
+    # Example for another product type
+    'labuus': {
+        'required': ['products'],
+        'optional': []
+    }
+}
