+++ constants/tcg/pokemon/scarletAndVioletEra/baseConfig.py
