+++ constants/tcg/pokemon/scarletAndVioletEra/blackBolt.py
