+++ constants/tcg/pokemon/scarletAndVioletEra/destinedRivals.py
