+++ constants/tcg/pokemon/scarletAndVioletEra/obsidianFlames.py
