+++ constants/tcg/pokemon/scarletAndVioletEra/paldeaEvolved.py
