+++ constants/tcg/pokemon/scarletAndVioletEra/paldeanFates.py
