+++ constants/tcg/pokemon/scarletAndVioletEra/paradoxRift.py
