+++ constants/tcg/pokemon/scarletAndVioletEra/scarletAndVioletBase.py
