+++ constants/tcg/pokemon/scarletAndVioletEra/setMap.py
