+++ constants/tcg/pokemon/scarletAndVioletEra/shroudedFable.py
