+++ constants/tcg/pokemon/scarletAndVioletEra/stellarCrown.py
