+++ constants/tcg/pokemon/scarletAndVioletEra/temporalForces.py
