+++ constants/tcg/pokemon/scarletAndVioletEra/whiteFlare.py
